Refactoring from Gramex
