--- config.ini
@@ -1,4 +1,0 @@
-[languagetool]
-hostname = http://localhost
-port = 8081
-apiversion = v2
