+++ nlg/grammar.py
@@ -1,26 +1,14 @@
 from inflect import engine
-from spacy.lang.en import LEMMA_INDEX, LEMMA_EXC, LEMMA_RULES
-from spacy.lemmatizer import Lemmatizer
 from tornado.template import Template
 
-from .utils import set_nlg_gramopt, load_spacy_model
-
-nlp = load_spacy_model()
+from nlg.utils import load_spacy_model, set_nlg_gramopt, get_lemmatizer
 
 infl = engine()
-L = Lemmatizer(LEMMA_INDEX, LEMMA_EXC, LEMMA_RULES)
-
-QUANT_FILTER_TOKENS = {
-    ">=": ["at least", "more than", "over"],
-    "<=": ["at most", "less than", "below"],
-    "==": ["of"],
-    "<": ["less than"],
-    ">": ["more than"],
-}
 
 
 def is_plural_noun(text):
-    doc = nlp(text)
+    """Whether given text is a plural noun."""
+    doc = load_spacy_model()(text)
     for t in list(doc)[::-1]:
         if not t.is_punct:
             return t.tag_ in ('NNS', 'NNPS')
@@ -30,8 +18,8 @@ def is_plural_noun(text):
 is_singular_noun = lambda x: not is_plural_noun(x)  # NOQA: E731
 
 
-@set_nlg_gramopt(source='G', fe_name="Concate Items")
-def concatenate_items(items, sep=", "):
+@set_nlg_gramopt(source='G', fe_name='Concate Items')
+def concatenate_items(items, sep=', '):
     """Concatenate a sequence of tokens into an English string.
 
     Parameters
@@ -51,15 +39,15 @@ def concatenate_items(items, sep=", "):
     if len(items) == 1:
         return items[0]
     items = list(map(str, items))
-    if sep == ", ":
+    if sep == ', ':
         s = sep.join(items[:-1])
-        s += " and " + items[-1]
+        s += ' and ' + items[-1]
     else:
         s = sep.join(items)
     return s
 
 
-@set_nlg_gramopt(source='G', fe_name="Pluralize")
+@set_nlg_gramopt(source='G', fe_name='Pluralize')
 def plural(word):
     """Pluralize a word.
 
@@ -79,16 +67,46 @@ def plural(word):
     return word
 
 
-@set_nlg_gramopt(source='G', fe_name="Singularize")
+@set_nlg_gramopt(source='G', fe_name='Singularize')
 def singular(word):
+    """
+    Singularize a word.
+
+    Parameters
+    ----------
+    word : str
+        Word to singularize.
+
+    Returns
+    -------
+    str
+        Singular of `word`.
+    """
+
     if is_plural_noun(word):
         word = infl.singular_noun(word)
     return word
 
 
-# @set_nlg_gramopt(source='G', fe_name="Pluralize by")
+# @set_nlg_gramopt(source='G', fe_name='Pluralize by')
 def pluralize_by(word, by):
-    """Pluralize a word depending on another argument."""
+    """
+    Pluralize a word depending on another argument.
+
+    Parameters
+    ----------
+    word : str
+        Word to pluralize
+    by : any
+        Any object checked for a pluralish value. If a sequence, it must have
+        length greater than 1 to qualify as plural.
+
+    Returns
+    -------
+    str
+        Plural or singular of `word`.
+    """
+
     if hasattr(by, '__iter__'):
         if len(by) > 1:
             word = plural(word)
@@ -102,53 +120,75 @@ def pluralize_by(word, by):
     return word
 
 
-# @set_nlg_gramopt(source='G', fe_name="Pluralize like")
+# @set_nlg_gramopt(source='G', fe_name='Pluralize like')
 def pluralize_like(x, y):
+    """
+    Pluralize a word if another is a plural.
+
+    Parameters
+    ----------
+    x : str
+        The word to pluralize.
+    y : str
+        The word to check.
+
+    Returns
+    -------
+    str
+        Plural of `x` if `y` is plural, else singular.
+    """
+
     if not is_plural_noun(y):
         return singular(x)
     return plural(x)
 
 
-@set_nlg_gramopt(source='str', fe_name="Capitalize")
+@set_nlg_gramopt(source='str', fe_name='Capitalize')
 def capitalize(word):
     return word.capitalize()
 
 
-@set_nlg_gramopt(source='str', fe_name="Lowercase")
+@set_nlg_gramopt(source='str', fe_name='Lowercase')
 def lower(word):
     return word.lower()
 
 
-@set_nlg_gramopt(source='str', fe_name="Swapcase")
+@set_nlg_gramopt(source='str', fe_name='Swapcase')
 def swapcase(word):
     return word.swapcase()
 
 
-@set_nlg_gramopt(source='str', fe_name="Title")
+@set_nlg_gramopt(source='str', fe_name='Title')
 def title(word):
     return word.title()
 
 
-@set_nlg_gramopt(source='str', fe_name="Uppercase")
+@set_nlg_gramopt(source='str', fe_name='Uppercase')
 def upper(word):
     return word.upper()
 
 
-# @set_nlg_gramopt(source="G", fe_name="Lemmatize")
+# @set_nlg_gramopt(source='G', fe_name='Lemmatize')
 def lemmatize(word, target_pos):
-    return L(word, target_pos)
+    return get_lemmatizer()(word, target_pos)
 
 
 def _token_inflections(x, y):
     """
-    Make changes in x lexically to turn it into y.
+    If two words share the same root, find lexical changes required for turning
+    one into another.
 
     Parameters
     ----------
-    x : [type]
-        [description]
-    y : [type]
-        [description]
+    x : spacy.token.Tokens
+    y : spacy.token.Tokens
+
+    Examples
+    --------
+    >>> _token_inflections('language', 'Language')
+    'upper'
+    >>> _token_inflections('language', 'languages')
+    'plural'
     """
     if x.lemma_ != y.lemma_:
         return False
@@ -172,6 +212,28 @@ def _token_inflections(x, y):
 
 
 def find_inflections(text, search, fh_args, df):
+    """
+    Find lexical inflections between words in input text and the search results
+    obtained from FormHandler arguments and dataframes.
+
+    Parameters
+    ----------
+    text : str
+        Input text
+    search : nlg.search.DFSearchResults
+        The DFSearchResults object corresponding to `text` and `df`
+    fh_args : dict
+        FormHandler arguments.
+    df : pandas.DataFrame
+        The source dataframe.
+
+    Returns
+    -------
+    dict
+        With keys as tokens found in the dataframe or FH args, and values as
+        list of inflections applied on them to make them closer match tokens in `text`.
+    """
+    nlp = load_spacy_model()
     text = nlp(text)
     inflections = {}
     for token, tklist in search.items():
