+++ nlg/search.py
@@ -1,4 +1,3 @@
-#! /usr/bin/env python
 # -*- coding: utf-8 -*-
 # vim:fenc=utf-8
 
@@ -10,34 +9,51 @@ from itertools import chain
 
 import numpy as np
 import pandas as pd
-from spacy import load
-from spacy.matcher import PhraseMatcher
+import six
 
 from nlg import utils
-from nlg.grammar import concatenate_items, find_inflections
-
-default_nlp = load("en_core_web_sm")
-matcher = utils.make_np_matcher(default_nlp)
+from nlg.grammar import find_inflections
 
 SEARCH_PRIORITIES = [
-    {'type': 'ne'},  # A match which is an NE gets the higest priority
-    {'location': 'fh_args'},  # then one that is a formhandler arg
-    {'location': 'colname'},  # then one that is a column name
+    {'type': 'ne'},  # A match which is a named entity gets the higest priority
+    {'location': 'fh_args'},  # than one that is a formhandler arg
+    {'location': 'colname'},  # than one that is a column name
     {'type': 'quant'},  # etc
     {'location': 'cell'}
 ]
 
 
 def _sort_search_results(items, priorities=SEARCH_PRIORITIES):
-    """Sort a list of search results by `priorities`."""
-    match_ix = [[p.items() <= item.items() for p in priorities] for item in items]
+    """
+    Sort a list of search results by `priorities`.
+
+    Parameters
+    ----------
+    items : dict
+        Dictionary containing search results, where keys are tokens and values
+        are lists of locations where the token was found. Preferably this should
+        be a `DFSearchResults` object.
+    priorities : list, optional
+        List of rules that allow sorting of search results. A `rule` is any
+        subset of a search result dictionary. Lower indices indicate higher priorities.
+
+    Returns
+    -------
+    dict
+        Prioritized search results - for each {token: search_matches} pair, sort
+        search_matches such that a higher priority search result is enabled.
+    """
+    match_ix = [[six.viewitems(p) <= six.viewitems(item) for p in priorities] for item in items]
     min_match = [m.index(True) for m in match_ix]
     items[min_match.index(min(min_match))]['enabled'] = True
     return items
 
 
 class DFSearchResults(dict):
-    """A convenience wrapper around `dict` to collect search results."""
+    """A convenience wrapper around `dict` to collect search results.
+
+    Different from `dict` in that values are always lists, and setting to
+    existing key appends to the list."""
 
     def __setitem__(self, key, value):
         if key not in self:
@@ -57,25 +73,57 @@ class DFSearchResults(dict):
         # unoverlap the keys
         to_remove = []
         for k in self:
-            if any([k in c for c in self.keys() - {k}]):
+            if any([k in c for c in six.viewkeys(self) - {k}]):
                 to_remove.append(k)
         for i in to_remove:
             del self[i]
 
 
 class DFSearch(object):
+    """Make a dataframe searchable."""
 
-    def __init__(self, df, nlp=default_nlp, **kwargs):
+    def __init__(self, df, nlp=None, **kwargs):
+        """Default constrictor.
+
+        Parameters
+        ----------
+        df : pd.DataFrame
+            The dataframe to search.
+        nlp : A `spacy.lang` model, optional
+        """
         self.df = df
         # What do results contain?
-        # A map of tokens to pandas slices
+        # A map of tokens to list of search results.
         self.results = DFSearchResults()
+        if not nlp:
+            nlp = utils.load_spacy_model()
         self.nlp = nlp
+        self.matcher = kwargs.get('matcher', utils.make_np_matcher(self.nlp))
 
-    def search(self, text, colname_fmt="df.columns[{}]",
-               cell_fmt="df['{}'].iloc[{}]", **kwargs):
+    def search(self, text, colname_fmt='df.columns[{}]',
+               cell_fmt='df["{}"].iloc[{}]', **kwargs):
+        """
+        Search the dataframe.
+
+        Parameters
+        ----------
+        text : str
+            The text to search.
+        colname_fmt : str, optional
+            String format to describe dataframe columns in the search results,
+            can be one of 'df.columns[{}]' or 'df[{}]'.
+        cell_fmt : str, optional
+            String format to describe dataframe values in the search results.
+            Can be one of 'df.iloc[{}, {}]', 'df.loc[{}, {}]', 'df[{}][{}]', etc.
+
+        Returns
+        -------
+        dict
+            A dictionary who's keys are tokens from `text` found in
+            the source dataframe, and values are a list of locations in the df
+            where they are found.
+        """
         self.search_nes(text)
-
         for token, ix in self.search_columns(text, **kwargs).items():
             ix = utils.sanitize_indices(self.df.shape, ix, 1)
             self.results[token] = {'location': 'colname', 'tmpl': colname_fmt.format(ix),
@@ -85,37 +133,56 @@ class DFSearch(object):
             x = utils.sanitize_indices(self.df.shape, x, 0)
             y = utils.sanitize_indices(self.df.shape, y, 1)
             self.results[token] = {
-                'location': "cell", 'tmpl': cell_fmt.format(self.df.columns[y], x),
+                'location': 'cell', 'tmpl': cell_fmt.format(self.df.columns[y], x),
                 'type': 'token'}
         self.search_quant([c.text for c in self.doc if c.pos_ == 'NUM'])
         return self.results
 
-    def search_nes(self, text, colname_fmt="df.columns[{}]", cell_fmt="df['{}'].iloc[{}]"):
+    def search_nes(self, text, colname_fmt='df.columns[{}]', cell_fmt='df["{}"].iloc[{}]'):
+        """Find named entities in text, and search for them in the dataframe.
+
+        Parameters
+        ----------
+        text : str
+            The text to search.
+        """
         self.doc = self.nlp(text)
-        self.ents = utils.ner(self.doc, matcher)
+        self.ents = utils.ner(self.doc, self.matcher)
         ents = [c.text for c in self.ents]
         for token, ix in self.search_columns(ents, literal=True).items():
             ix = utils.sanitize_indices(self.df.shape, ix, 1)
             self.results[token] = {
-                'location': "colname",
+                'location': 'colname',
                 'tmpl': colname_fmt.format(ix), 'type': 'ne'
             }
         for token, (x, y) in self.search_table(ents, literal=True).items():
             x = utils.sanitize_indices(self.df.shape, x, 0)
             y = utils.sanitize_indices(self.df.shape, y, 1)
             self.results[token] = {
-                'location': "cell",
+                'location': 'cell',
                 'tmpl': cell_fmt.format(self.df.columns[y], x), 'type': 'ne'}
 
     def search_table(self, text, **kwargs):
-        kwargs['array'] = self.df.copy()  # ToDo: Is this necessary?
+        """Search the `.values` attribute of the dataframe for tokens in `text`."""
+        kwargs['array'] = self.df.copy()
         return self._search_array(text, **kwargs)
 
     def search_columns(self, text, **kwargs):
+        """Search df columns for tokens in `text`."""
         kwargs['array'] = self.df.columns
         return self._search_array(text, **kwargs)
 
-    def search_quant(self, quants, nround=2, cell_fmt="df['{}'].iloc[{}]"):
+    def search_quant(self, quants, nround=2, cell_fmt='df["{}"].iloc[{}]'):
+        """Search the dataframe for a set of quantitative values.
+
+        Parameters
+        ----------
+        quants : list / array like
+            The values to search.
+        nround : int, optional
+            Numeric values in the dataframe are rounded to these many
+            significant digits before searching.
+        """
         dfclean = utils.sanitize_df(self.df, nround)
         quants = np.array(quants)
         n_quant = quants.astype('float').round(2)
@@ -124,28 +191,63 @@ class DFSearch(object):
             y = utils.sanitize_indices(dfclean.shape, y, 1)
             tk = quants[n_quant == dfclean.iloc[x, y]][0].item()
             self.results[tk] = {
-                'location': "cell", 'tmpl': cell_fmt.format(self.df.columns[y], x),
+                'location': 'cell', 'tmpl': cell_fmt.format(self.df.columns[y], x),
                 'type': 'quant'}
 
     def _search_array(self, text, array, literal=False,
                       case=False, lemmatize=True, nround=2):
-        """Search for tokens in text within a pandas array.
-        Return {token: array_int_index}"""
+        """Search for tokens in text within an array.
+
+        Parameters
+        ----------
+        text : str or spacy document
+            Text to search
+        array : array-like
+            Array to search in.
+        literal : bool, optional
+            Whether to match tokens to values literally.
+        case : bool, optional
+            If true, run a case sensitive search.
+        lemmatize : bool, optional
+            If true (default), search on lemmas of tokens and values.
+        nround : int, optional
+            Significant digits used to round `array` before searching.
+
+        Returns
+        -------
+        dict
+            Mapping of tokens to a sequence of indices within `array`.
+
+        Example
+        -------
+        >>> _search_array('3', np.arange(5))
+        {'3': [2]}
+        >>> df = pd.DataFrame(np.eye(3), columns='one punch man'.split())
+        >>> _search_array('1', df.values)
+        {'1': [(0, 0), (1, 1), (2, 2)]}
+        >>> _search_array('punched man', df.columns)
+        {'punched': [1], 'man': [2]}
+        >>> _search_array('1 2 buckle my shoe', df.index)
+        {'1': [1], '2': [2]}
+        """
         if literal:
             # Expect text to be a list of strings, no preprocessing on anything.
             if not isinstance(text, list):
                 raise TypeError('text is expected to be list of strs when literal=True.')
-            if not set([type(c) for c in text]).issubset({str, float, int}):
+            valid_types = {float, int, six.text_type}
+            if not set([type(c) for c in text]).issubset(valid_types):
                 raise TypeError('text can contain only strings or numbers when literal=True.')
             tokens = {c: str(c) for c in text}
         elif lemmatize:
             tokens = {c.lemma_: c.text for c in self.nlp(text)}
             if array.ndim == 1:
+                array = [c if isinstance(c, six.text_type) else six.u(c) for c in array]
                 array = [self.nlp(c) for c in array]
                 array = pd.Series([token.lemma_ for doc in array for token in doc])
             else:
                 for col in array.columns[array.dtypes == np.dtype('O')]:
-                    s = [self.nlp(c) for c in array[col]]
+                    s = [c if isinstance(c, six.text_type) else six.u(c) for c in array[col]]
+                    s = [self.nlp(c) for c in s]
                     try:
                         array[col] = [token.lemma_ for doc in s for token in doc]
                     except ValueError:
@@ -175,70 +277,46 @@ class DFSearch(object):
         return {}
 
 
-def search_concatenations(text, df):
-    doc = default_nlp(text)
-    matcher = PhraseMatcher(default_nlp.vocab)
-    patterns = []
-    for _, series in df.items():
-        if series.dtype == np.dtype('O'):
-            patterns.extend([default_nlp(x) for x in series])
-    matcher.add("cell", None, *patterns)
-    spans = []
-    for _, start, end in matcher(doc):
-        spans.append(doc[start:end].text)
-    ideal = concatenate_items(spans)
-    if ideal not in text:
-        return ''
-    mask = df.isin(spans)
-    if mask.sum().sum() < 2:
-        return ''
-    # search for columns:
-    col = df.columns[mask.any(0)][0]
-    y = mask[col].nonzero()[0]
-    if set(np.diff(y)) == {1}:
-        return {ideal: "df.iloc[{}:{}, '{}']".format(y.min(), y.max(), col)}
-
-
-def lemmatized_df_search(x, y, fmt_string="df.columns[{}]"):
-    search_res = {}
-    tokens = list(chain(*x))
-    colnames = list(chain(*[default_nlp(c) for c in y]))
-    for i, xx in enumerate(colnames):
-        for yy in tokens:
-            if xx.lemma_ == yy.lemma_:
-                search_res[yy.text] = fmt_string.format(i)
-    return search_res
-
-
-def search_args(entities, args, lemmatized=True, fmt="fh_args['{}'][{}]",
-                argkeys=('_sort', '_by')):
+def search_args(entities, args, lemmatized=True, fmt='fh_args["{}"][{}]',
+                argkeys=('_sort', '_by', '_c')):
     """
-    Search formhandler arguments.
+    Search formhandler arguments, as parsed by g1, for a set of tokens.
 
     Parameters
     ----------
     entities : list
-        list of spacyy entities
-    args : Formhandler args
-        [description]
+        list of named entities found in the source text
+    args : dict
+        FormHandler args as parsed by g1.url.parse(...).searchList
     lemmatized : bool, optional
-        whether to lemmatize search (the default is True, which [default_description])
+        whether to search on lemmas of text values
     fmt : str, optional
-        format used in the template (the default is "args['{}'][{}]", which [default_description])
+        String format used to describe FormHandler arguments in the template
     argkeys : list, optional
-        keys to be considered for the search (the default is None, which [default_description])
+        Formhandler argument keys to be considered for the search. Any key not
+        present in this will be ignored.
+        # TODO: Column names can be keys too!!
 
     Returns
     -------
-    [type]
-        [description]
+    dict
+        Mapping of entities / tokens to objects describing where they are found
+        in Formhandler arguemnts. Each search result object has the following
+        structure:
+        {
+            'type': 'some token',
+            'location': 'fh_args',
+            'tmpl': 'fh_args['_by'][0]'  # The template that gets this token from fh_args
+        }
     """
+    nlp = utils.load_spacy_model()
     args = {k: v for k, v in args.items() if k in argkeys}
     search_res = {}
     ent_tokens = list(chain(*entities))
     for k, v in args.items():
-        # argtokens = list(chain(*[re.findall(r"\w+", f) for f in v]))
-        argtokens = list(chain(*[default_nlp(c) for c in v]))
+        v = [t.lstrip('-') for t in v]
+        # argtokens = list(chain(*[re.findall(r'\w+', f) for f in v]))
+        argtokens = list(chain(*[nlp(c) for c in v]))
         for i, x in enumerate(argtokens):
             for y in ent_tokens:
                 if lemmatized:
@@ -254,59 +332,33 @@ def search_args(entities, args, lemmatized=True, fmt="fh_args['{}'][{}]",
     return search_res
 
 
-def search_df(tokens, df):
-    """Search a dataframe for tokens and return the coordinates."""
-    search_res = {}
-    txt_tokens = np.array([c.text for c in tokens])
-    coltype = df.columns.dtype
-    ixtype = df.index.dtype
-
-    # search in columns
-    column_ix = np.arange(df.shape[1])[df.columns.astype(str).isin(txt_tokens)]
-    for ix in column_ix:
-        token = df.columns[ix]
-        ix = utils.sanitize_indices(df.shape, ix, 1)
-        search_res[token] = "df.columns[{}]".format(ix)
-
-    # search in index
-    index_ix = df.index.astype(str).isin(txt_tokens)
-    for token in df.index[index_ix]:
-        if token not in search_res:
-            if ixtype == np.dtype("O"):
-                indexer = "df.loc['{}']".format(token)
-            else:
-                indexer = "df.loc[{}]".format(token)
-            search_res[token] = indexer
-
-    # search in table
-    for token in txt_tokens:
-        if token not in search_res:
-            mask = df.values.astype(str) == token
-            try:
-                column = df.columns[mask.sum(0).astype(bool)][0]
-                # don't sanitize column
-                index = df.index[mask.sum(1).astype(bool)][0]
-                index = utils.sanitize_indices(df.shape, index, 0)
-            except IndexError:
-                continue
-            if coltype == np.dtype("O"):
-                col_indexer = "'{}'".format(column)
-            else:
-                col_indexer = str(column)
-            if ixtype == np.dtype("O"):
-                ix_indexer = "'{}'".format(index)
-            else:
-                ix_indexer = str(index)
-            search_res[token] = "df.iloc[{}][{}]".format(ix_indexer, col_indexer)
-
-    unfound = [token for token in tokens if token.text not in search_res]
-    search_res.update(lemmatized_df_search(unfound, df.columns))
-    return search_res
-
-
 def templatize(text, args, df):
-    """Process a piece of text and templatize it according to a dataframe."""
-    df.fillna(value="", inplace=True)
+    """Construct a tornado template which regenerates some
+    text from a dataframe and formhandler arguments.
+
+    The pipeline consists of:
+    1. cleaning the text and the dataframe
+    2. searching the dataframe and FH args for tokens in the text
+    3. detecting inflections on the tokens.
+
+    Parameters
+    ----------
+    text : str
+        Input text
+    args : dict
+        Formhandler arguments
+    df : pd.DataFrame
+        Source dataframe.
+
+    Returns
+    --------
+    tuple
+        of search results, cleaned text and token inflections. The webapp uses
+        these to construct a tornado template.
+    """
+    text = six.u(text)
+    args = {six.u(k): [six.u(c) for c in v] for k, v in args.items()}
+    utils.load_spacy_model()
     clean_text = utils.sanitize_text(text)
     args = utils.sanitize_fh_args(args)
     dfs = DFSearch(df)
