+++ nlg/tests/test_search.py
@@ -11,8 +11,10 @@ import re
 import unittest
 
 import pandas as pd
+from tornado.template import Template
 
 from nlg import search, utils
+
 nlp = utils.load_spacy_model()
 matcher = utils.make_np_matcher(nlp)
 
@@ -44,7 +46,7 @@ class TestDFSearch(unittest.TestCase):
     def test_dfsearch_lemmatized(self):
         df = pd.DataFrame.from_dict(
             {
-                "partner": ["Lata", "Asha", "Rafi"],
+                "partner": ["Lata Mangeshkar", "Asha Bhosale", "Mohammad Rafi"],
                 "song": [20, 5, 15],
             }
         )
@@ -54,7 +56,8 @@ class TestDFSearch(unittest.TestCase):
             dfs.search(sent, lemmatize=True),
             {
                 'songs': [{"location": "colname", "type": "token", "tmpl": "df.columns[1]"}],
-                'Lata': [{'location': 'cell', 'tmpl': "df['partner'].iloc[0]", 'type': 'token'}],
+                'Lata Mangeshkar': [
+                    {'location': 'cell', 'tmpl': 'df["partner"].iloc[0]', 'type': 'ne'}],
             }
         )
 
@@ -69,10 +72,10 @@ class TestDFSearch(unittest.TestCase):
             dfs.search(sent),
             {
                 'Spencer Tracy': [
-                    {'location': 'cell', 'tmpl': "df['name'].iloc[0]", 'type': 'ne'}
+                    {'location': 'cell', 'tmpl': 'df["name"].iloc[0]', 'type': 'ne'}
                 ],
                 'voted': [{'location': 'colname', 'tmpl': 'df.columns[-1]', 'type': 'token'}],
-                'actor': [{'location': 'cell', 'tmpl': "df['category'].iloc[-4]", 'type': 'token'}]
+                'actor': [{'location': 'cell', 'tmpl': 'df["category"].iloc[-4]', 'type': 'token'}]
             }
         )
 
@@ -141,7 +144,10 @@ class TestSearch(unittest.TestCase):
             tmpl = [t for t in tmpls if t.get('enabled', False)][0]
             actual = actual.replace(token, '{{{{ {} }}}}'.format(tmpl['tmpl']))
         cleaner = lambda x: re.sub(r"\s+", " ", x)  # NOQA: E731
-        self.assertEqual(*map(cleaner, (ideal, actual)))
+        ideal, actual = map(cleaner, (ideal, actual))
+        ideal = Template(ideal).generate(df=df, fh_args=args)
+        actual = Template(actual).generate(df=df, fh_args=args)
+        self.assertEqual(ideal, actual)
         self.assertDictEqual(
             inflections,
             {
